DB Version 2.1

+++ CGUApp/src/cgu/edu/ist380/alsuhaibanyy/alghosona/db/BuildingCoordinate.java
@@ -2,19 +2,19 @@ package cgu.edu.ist380.alsuhaibanyy.alghosona.db;
 
 public class BuildingCoordinate {
 	
-	private long mLongitude;
-	private long mLatitude;
+	private double mLongitude;
+	private double mLatitude;
 	private int mBuildingID;
-	public long getmLongitude() {
+	public double getmLongitude() {
 		return mLongitude;
 	}
-	public void setmLongitude(long mLongitude) {
+	public void setmLongitude(double mLongitude) {
 		this.mLongitude = mLongitude;
 	}
-	public long getmLatitude() {
+	public double getmLatitude() {
 		return mLatitude;
 	}
-	public void setmLatitude(long mLatitude) {
+	public void setmLatitude(double mLatitude) {
 		this.mLatitude = mLatitude;
 	}
 	public int getmBuildingID() {
